Fix Selecting Games

+++ Sudachi/Sudachi/Core/core/hle/kernel/k_page_table_base.cpp
@@ -92,12 +92,12 @@ bool Shouldu32oru8() {
     return GetBoolFromUserDefaults("ClearBackingRegion");
 }
 
-void ClearBackingRegion(Core::System& system, KPhysicalAddress addr, u64 size, u32 fill_value) {
+void ClearBackingRegion(Core::System& system, KPhysicalAddress addr, uint_fast8_t size, uint8_t fill_value) {
         system.DeviceMemory().buffer.ClearBackingRegion(GetInteger(addr) - Core::DramMemoryMap::Base,
                                                         size, fill_value);
 }
 
-void ClearBackingRegion2(Core::System& system, KPhysicalAddress addr, uint_fast8_t size, uint8_t fill_value) {
+void ClearBackingRegion2(Core::System& system, KPhysicalAddress addr, u64 size, u32 fill_value) {
         system.DeviceMemory().buffer.ClearBackingRegion(GetInteger(addr) - Core::DramMemoryMap::Base,
                                                         size, fill_value);
 }
